# Vectorized Multiply (on Avx512)