--- a/src/Nethermind.Int256/UInt256.cs
+++ b/src/Nethermind.Int256/UInt256.cs
@@ -1010,25 +1010,142 @@ namespace Nethermind.Int256
                 Unsafe.AsRef(in res.u1) = high;
                 return;
             }
+            
+            if (!Avx512F.IsSupported || !Avx512DQ.IsSupported)
+            {
+                ref ulong rx = ref Unsafe.As<UInt256, ulong>(ref Unsafe.AsRef(in x));
+                ref ulong ry = ref Unsafe.As<UInt256, ulong>(ref Unsafe.AsRef(in y));
 
-            ref ulong rx = ref Unsafe.As<UInt256, ulong>(ref Unsafe.AsRef(in x));
-            ref ulong ry = ref Unsafe.As<UInt256, ulong>(ref Unsafe.AsRef(in y));
+                (ulong carry, ulong r0) = Multiply64(rx, ry);
+                UmulHop(carry, Unsafe.Add(ref rx, 1), ry, out carry, out ulong res1);
+                UmulHop(carry, Unsafe.Add(ref rx, 2), ry, out carry, out ulong res2);
+                ulong res3 = Unsafe.Add(ref rx, 3) * ry + carry;
 
-            (ulong carry, ulong r0) = Multiply64(rx, ry);
-            UmulHop(carry, Unsafe.Add(ref rx, 1), ry, out carry, out ulong res1);
-            UmulHop(carry, Unsafe.Add(ref rx, 2), ry, out carry, out ulong res2);
-            ulong res3 = Unsafe.Add(ref rx, 3) * ry + carry;
+                UmulHop(res1, rx, Unsafe.Add(ref ry, 1), out carry, out ulong r1);
+                UmulStep(res2, Unsafe.Add(ref rx, 1), Unsafe.Add(ref ry, 1), carry, out carry, out res2);
+                res3 = res3 + Unsafe.Add(ref rx, 2) * Unsafe.Add(ref ry, 1) + carry;
 
-            UmulHop(res1, rx, Unsafe.Add(ref ry, 1), out carry, out ulong r1);
-            UmulStep(res2, Unsafe.Add(ref rx, 1), Unsafe.Add(ref ry, 1), carry, out carry, out res2);
-            res3 = res3 + Unsafe.Add(ref rx, 2) * Unsafe.Add(ref ry, 1) + carry;
+                UmulHop(res2, rx, Unsafe.Add(ref ry, 2), out carry, out ulong r2);
+                res3 = res3 + Unsafe.Add(ref rx, 1) * Unsafe.Add(ref ry, 2) + carry;
 
-            UmulHop(res2, rx, Unsafe.Add(ref ry, 2), out carry, out ulong r2);
-            res3 = res3 + Unsafe.Add(ref rx, 1) * Unsafe.Add(ref ry, 2) + carry;
+                ulong r3 = res3 + rx * Unsafe.Add(ref ry, 3);
 
-            ulong r3 = res3 + rx * Unsafe.Add(ref ry, 3);
+                res = new UInt256(r0, r1, r2, r3);
+            }
+            else
+            {
 
-            res = new UInt256(r0, r1, r2, r3);
+                // Unpack the four 64-bit limbs (little-endian: u0 is least-significant)
+                ulong a0 = x.u0, a1 = x.u1, a2 = x.u2, a3 = x.u3;
+                ulong b0 = y.u0, b1 = y.u1, b2 = y.u2, b3 = y.u3;
+
+                // --- Compute the 10 64x64–bit products using our vectorized method ---
+
+                // Group 1: 8 products
+                Vector512<ulong> vecA1 = Vector512.Create(a0, a0, a1, a0, a1, a2, a0, a1);
+                Vector512<ulong> vecB1 = Vector512.Create(b0, b1, b0, b2, b1, b0, b3, b2);
+                Mul64Vector(vecA1, vecB1, out Vector512<ulong> lo1, out Vector512<ulong> hi1);
+
+                // Extract products from group1
+                ulong P00_lo = lo1.GetElement(0), P00_hi = hi1.GetElement(0);
+                ulong P01_lo = lo1.GetElement(1), P01_hi = hi1.GetElement(1);
+                ulong P10_lo = lo1.GetElement(2), P10_hi = hi1.GetElement(2);
+                ulong P02_lo = lo1.GetElement(3), P02_hi = hi1.GetElement(3);
+                ulong P11_lo = lo1.GetElement(4), P11_hi = hi1.GetElement(4);
+                ulong P20_lo = lo1.GetElement(5), P20_hi = hi1.GetElement(5);
+                ulong P03_lo = lo1.GetElement(6), P03_hi = hi1.GetElement(6);
+                ulong P12_lo = lo1.GetElement(7), P12_hi = hi1.GetElement(7);
+
+                // Group 2: 2 products
+                Vector512<ulong> vecA2 = Vector512.Create(a2, a3, 0UL, 0UL, 0UL, 0UL, 0UL, 0UL);
+                Vector512<ulong> vecB2 = Vector512.Create(b1, b0, 0UL, 0UL, 0UL, 0UL, 0UL, 0UL);
+                Mul64Vector(vecA2, vecB2, out Vector512<ulong> lo2, out Vector512<ulong> hi2);
+                ulong P21_lo = lo2.GetElement(0); // P21_hi is not needed (contributes only above 256 bits)
+                ulong P30_lo = lo2.GetElement(1); // Likewise for P30_hi
+
+                // --- Package each 128-bit partial product into a UInt256 (with proper shifting) ---
+                // (Recall: a 128–bit product is given as (lo, hi), where lo is the lower 64 bits and hi the upper 64 bits.)
+
+                // P00 (no shift)
+                UInt256 part0 = new UInt256(P00_lo, P00_hi, 0, 0);
+
+                // P01 and P10 (each shifted left by 64 bits)
+                UInt256 part64a = new UInt256(0, P01_lo, P01_hi, 0);
+                UInt256 part64b = new UInt256(0, P10_lo, P10_hi, 0);
+                UInt256 sum64;
+                AddImpl(part64a, part64b, out sum64);
+
+                // P02, P11 and P20 (each shifted left by 128 bits)
+                UInt256 part128a = new UInt256(0, 0, P02_lo, P02_hi);
+                UInt256 part128b = new UInt256(0, 0, P11_lo, P11_hi);
+                UInt256 part128c = new UInt256(0, 0, P20_lo, P20_hi);
+                UInt256 sum128, temp;
+                AddImpl(part128a, part128b, out temp);
+                AddImpl(temp, part128c, out sum128);
+
+                // P03, P12, P21 and P30 (shifted left by 192 bits – note only the low 64 bits matter)
+                UInt256 part192a = new UInt256(0, 0, 0, P03_lo);
+                UInt256 part192b = new UInt256(0, 0, 0, P12_lo);
+                UInt256 part192c = new UInt256(0, 0, 0, P21_lo);
+                UInt256 part192d = new UInt256(0, 0, 0, P30_lo);
+                UInt256 sum192;
+                AddImpl(part192a, part192b, out temp);
+                AddImpl(temp, part192c, out temp);
+                AddImpl(temp, part192d, out sum192);
+
+                // --- Sum all the partial products using AddImpl ---
+                UInt256 intermediate;
+                AddImpl(part0, sum64, out intermediate);
+                AddImpl(intermediate, sum128, out intermediate);
+                AddImpl(intermediate, sum192, out res);
+            }
+        }
+
+        
+        // Vectorized 64x64 multiply: given vectors 'a' and 'b' (each 8 lanes),
+        // computes per lane:
+        //   product = a * b = (hi, lo)
+        // using the splitting method since there is no MultiplyHigh intrinsic.
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static void Mul64Vector(Vector512<ulong> a, Vector512<ulong> b,
+                                          out Vector512<ulong> lo, out Vector512<ulong> hi)
+        {
+            // Mask for the lower 32 bits.
+            Vector512<ulong> mask32 = Vector512.Create(0xFFFFFFFFUL);
+
+            // Split each 64-bit operand into 32-bit halves:
+            // a0 = lower 32 bits, a1 = upper 32 bits
+            Vector512<ulong> a0 = Avx512F.And(a, mask32);
+            Vector512<ulong> a1 = Avx512F.ShiftRightLogical(a, 32);
+            Vector512<ulong> b0 = Avx512F.And(b, mask32);
+            Vector512<ulong> b1 = Avx512F.ShiftRightLogical(b, 32);
+
+            // Compute the four 32x32 partial products.
+            // Each multiplication here is on 32-bit values, so the result fits in 64 bits.
+            Vector512<ulong> u0 = Avx512DQ.MultiplyLow(a0, b0); // a0 * b0
+            Vector512<ulong> u1 = Avx512DQ.MultiplyLow(a0, b1); // a0 * b1
+            Vector512<ulong> u2 = Avx512DQ.MultiplyLow(a1, b0); // a1 * b0
+            Vector512<ulong> u3 = Avx512DQ.MultiplyLow(a1, b1); // a1 * b1
+
+            // Now, compute t = (u0 >> 32) + (u1 & mask32) + (u2 & mask32)
+            Vector512<ulong> u0_hi = Avx512F.ShiftRightLogical(u0, 32);
+            Vector512<ulong> u1_lo = Avx512F.And(u1, mask32);
+            Vector512<ulong> u2_lo = Avx512F.And(u2, mask32);
+            Vector512<ulong> t = Avx512F.Add(Avx512F.Add(u0_hi, u1_lo), u2_lo);
+
+            // The extra carry: c = t >> 32.
+            Vector512<ulong> c = Avx512F.ShiftRightLogical(t, 32);
+
+            // Now, assemble the lower 64 bits:
+            // low part of u0 is u0 & mask32; low 32 bits of t are (t & mask32) shifted left 32.
+            Vector512<ulong> u0_lo = Avx512F.And(u0, mask32);
+            Vector512<ulong> t_lo = Avx512F.And(t, mask32);
+            lo = Avx512F.Or(u0_lo, Avx512F.ShiftLeftLogical(t_lo, 32));
+
+            // The high 64 bits are: u3 + (u1 >> 32) + (u2 >> 32) + c.
+            Vector512<ulong> u1_hi = Avx512F.ShiftRightLogical(u1, 32);
+            Vector512<ulong> u2_hi = Avx512F.ShiftRightLogical(u2, 32);
+            hi = Avx512F.Add(Avx512F.Add(Avx512F.Add(u3, u1_hi), u2_hi), c);
         }
 
         public void Multiply(in UInt256 a, out UInt256 res) => Multiply(this, a, out res);